+++ src/index.html
@@ -5,7 +5,7 @@
     <title>Data Lovers</title>
     <link rel="stylesheet" href="style.css" />
     <div id="root"></div>
-    <script src="main.js" type="module"></script>
+    <script src="data.js" type="module"></script>
   </head>
   <body>
     <header>
@@ -23,9 +23,17 @@
           <li><a href="#">MORE</a></li>
         </ul>
       </nav>
+      <select class="Sort">
+        <option value= "1" id="ascendente">Sort A-Z</option>
+        <option value= "2" id="descendente">Sort Z-A</option>
+
+
+    </select>
     </header>
     <main> 
       <div class="contenedor"></div>
+
+     
     </main>
   </body>
 </html>
