+++ src/style.css
@@ -256,3 +256,9 @@ input[type="text"]{
   border-radius: 12px;
   
 }
+.Sort{
+
+  color: black;
+  top: 300px;
+  
+}
